create sub-directories within ditributedTest in GeodeRedisModule
