+++ geode-redis/src/distributedTest/java/org/apache/geode/redis/executors/general/ExistsDUnitTest.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.redis.executors;
+package org.apache.geode.redis.executors.general;
 
 import static org.assertj.core.api.Assertions.assertThat;
